[PIR] add print function for pd_op.while
