+++ paddle/fluid/framework/new_executor/pir_adaptor/pir_adaptor_util.cc
@@ -43,6 +43,7 @@
 #include "glog/logging.h"
 #include "paddle/fluid/framework/op_info.h"
 #include "paddle/fluid/framework/operator.h"
+#include "paddle/fluid/pir/dialect/operator/ir/control_flow_op.h"
 #include "paddle/fluid/pir/dialect/operator/ir/manual_op.h"
 
 namespace paddle {
