+++ test/cpp/pir/control_flow_dialect/if_op_test.cc
@@ -14,7 +14,7 @@
 #include <gtest/gtest.h>
 #include <iostream>
 
-#include "paddle/fluid/pir/dialect/operator/ir/manual_op.h"
+#include "paddle/fluid/pir/dialect/operator/ir/control_flow_op.h"
 #include "paddle/fluid/pir/dialect/operator/ir/op_dialect.h"
 #include "paddle/fluid/pir/dialect/operator/ir/pd_op.h"
 #include "paddle/pir/core/builder.h"
